# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oxymed Medikal API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
}

export interface AuthTokenResult {
  token: string;
  user: AdminUser;
}

export interface Slider {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  ctaPrimaryText?: string | null;
  /** @nullable */
  ctaPrimaryHref?: string | null;
  /** @nullable */
  ctaSecondaryText?: string | null;
  /** @nullable */
  ctaSecondaryHref?: string | null;
  sortOrder: number;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface SliderInput {
  title: string;
  subtitle?: string;
  description?: string;
  imageUrl?: string;
  ctaPrimaryText?: string;
  ctaPrimaryHref?: string;
  ctaSecondaryText?: string;
  ctaSecondaryHref?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export interface SliderUpdate {
  title?: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  ctaPrimaryText?: string | null;
  /** @nullable */
  ctaPrimaryHref?: string | null;
  /** @nullable */
  ctaSecondaryText?: string | null;
  /** @nullable */
  ctaSecondaryHref?: string | null;
  sortOrder?: number;
  isActive?: boolean;
}

export interface ProductCategory {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProductCategoryInput {
  name: string;
  slug: string;
  description?: string;
  sortOrder?: number;
}

export interface ProductCategoryUpdate {
  name?: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  sortOrder?: number;
}

export interface ProductSpec {
  label: string;
  value: string;
}

export interface Product {
  id: number;
  /** @nullable */
  categoryId?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  specs?: ProductSpec[];
  sortOrder: number;
  published: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  categoryId?: number;
  title: string;
  description?: string;
  imageUrl?: string;
  specs?: ProductSpec[];
  sortOrder?: number;
  published?: boolean;
}

export interface ProductUpdate {
  /** @nullable */
  categoryId?: number | null;
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  specs?: ProductSpec[];
  sortOrder?: number;
  published?: boolean;
}

export interface ProductListResponse {
  items: Product[];
  total: number;
}

export interface NewsItem {
  id: number;
  title: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  content?: string | null;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  slug: string;
  published: boolean;
  publishedAt: string;
  createdAt: string;
  updatedAt: string;
}

export interface NewsInput {
  title: string;
  excerpt?: string;
  content?: string;
  category?: string;
  imageUrl?: string;
  slug: string;
  published?: boolean;
  publishedAt?: string;
}

export interface NewsUpdate {
  title?: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  content?: string | null;
  category?: string;
  /** @nullable */
  imageUrl?: string | null;
  slug?: string;
  published?: boolean;
  publishedAt?: string;
}

export interface NewsListResponse {
  items: NewsItem[];
  total: number;
}

export interface ReferenceItem {
  id: number;
  title: string;
  projectType: string;
  /** @nullable */
  capacity?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  createdAt: string;
  updatedAt: string;
}

export interface ReferenceInput {
  title: string;
  projectType: string;
  capacity?: string;
  city?: string;
  imageUrl?: string;
  category?: string;
}

export interface ReferenceUpdate {
  title?: string;
  projectType?: string;
  /** @nullable */
  capacity?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  category?: string;
}

export interface ReferenceListResponse {
  items: ReferenceItem[];
  total: number;
}

export interface QuoteItem {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  projectType?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  applicationArea?: string | null;
  /** @nullable */
  notes?: string | null;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface QuoteInput {
  fullName: string;
  email: string;
  phone: string;
  company?: string;
  jobTitle?: string;
  projectType?: string;
  city?: string;
  applicationArea?: string;
  notes?: string;
}

export interface QuoteStatusUpdate {
  status: string;
}

export interface QuoteListResponse {
  items: QuoteItem[];
  total: number;
}

export interface CorporateSection {
  id: number;
  sectionKey: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CorporateSectionInput {
  title?: string;
  subtitle?: string;
  content?: string;
  imageUrl?: string;
}

export interface Setting {
  id: number;
  settingKey: string;
  settingValue: string;
}

export interface SettingInput {
  settingValue: string;
}

export interface SettingsMap {[key: string]: string}

export interface MediaFile {
  id: number;
  filename: string;
  objectPath: string;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  size?: number | null;
  /** @nullable */
  alt?: string | null;
  createdAt: string;
}

export interface MediaUploadInput {
  filename: string;
  objectPath: string;
  mimeType?: string;
  size?: number;
  alt?: string;
}

export interface MediaListResponse {
  items: MediaFile[];
  total: number;
}

export interface PresignedUrlInput {
  name: string;
  size: number;
  contentType: string;
}

export interface PresignedUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface DashboardStats {
  products: number;
  news: number;
  references: number;
  pendingQuotes: number;
  totalQuotes: number;
  sliders: number;
}

export type ListSlidersParams = {
activeOnly?: boolean;
};

export type ListProductsParams = {
categoryId?: number;
published?: boolean;
page?: number;
limit?: number;
};

export type ListNewsParams = {
category?: string;
published?: boolean;
page?: number;
limit?: number;
};

export type ListReferencesParams = {
category?: string;
page?: number;
limit?: number;
};

export type ListQuotesParams = {
status?: string;
page?: number;
limit?: number;
};

export type ListMediaFilesParams = {
page?: number;
limit?: number;
};