# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Oxymed Medikal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: adminMe
      tags: [auth]
      summary: Get current admin user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current admin user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # SLIDERS
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news posts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: published
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of news
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsListResponse"
    post:
      operationId: createNews
      tags: [news]
      summary: Create news post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
  /news/{id}:
    get:
      operationId: getNewsItem
      tags: [news]
      summary: Get news by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateNews
      tags: [news]
      summary: Update news post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete news post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # REFERENCES
...
  /corporate/{sectionKey}:
    get:
      operationId: getCorporateSection
      tags: [corporate]
      summary: Get corporate section by key
      parameters:
        - name: sectionKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Corporate section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorporateSection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertCorporateSection
      tags: [corporate]
      summary: Create or update corporate section
      security:
        - bearerAuth: []
      parameters:
        - name: sectionKey
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CorporateSectionInput"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorporateSection"

  # SETTINGS
...
  /media:
    get:
      operationId: listMediaFiles
      tags: [media]
      summary: List media files (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of media files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaListResponse"
  /media/upload:
    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload a media file (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUploadInput"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFile"
  /media/request-upload-url:
    post:
      operationId: requestMediaUploadUrl
      tags: [media]
      summary: Request presigned upload URL for a media file
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresignedUrlInput"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignedUrlResponse"
  /media/{id}:
    delete:
      operationId: deleteMediaFile
      tags: [media]
      summary: Delete a media file
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # DASHBOARD STATS
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [auth]
      summary: Get dashboard statistics (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # AUTH
    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthTokenResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
    AdminUser:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string

    # SLIDERS
...
    NewsItem:
      type: object
      required: [id, title, category, slug, published, publishedAt, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        slug:
          type: string
        published:
          type: boolean
        publishedAt:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    NewsInput:
      type: object
      required: [title, slug]
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        slug:
          type: string
        published:
          type: boolean
        publishedAt:
          type: string
    NewsUpdate:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        slug:
          type: string
        published:
          type: boolean
        publishedAt:
          type: string
    NewsListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NewsItem"
        total:
          type: integer

    # REFERENCES
...
    CorporateSection:
      type: object
      required: [id, sectionKey, createdAt, updatedAt]
      properties:
        id:
          type: integer
        sectionKey:
          type: string
        title:
          type: ["string", "null"]
        subtitle:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CorporateSectionInput:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        content:
          type: string
        imageUrl:
          type: string

    # SETTINGS
...
    MediaFile:
      type: object
      required: [id, filename, objectPath, createdAt]
      properties:
        id:
          type: integer
        filename:
          type: string
        objectPath:
          type: string
        mimeType:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        alt:
          type: ["string", "null"]
        createdAt:
          type: string
    MediaUploadInput:
      type: object
      required: [filename, objectPath]
      properties:
        filename:
          type: string
        objectPath:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        alt:
          type: string
    MediaListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MediaFile"
        total:
          type: integer
    PresignedUrlInput:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
    PresignedUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string

    # DASHBOARD
    DashboardStats:
      type: object
      required: [products, news, references, pendingQuotes, totalQuotes, sliders]
      properties:
        products:
          type: integer
        news:
          type: integer
        references:
          type: integer
        pendingQuotes:
          type: integer
        totalQuotes:
          type: integer
        sliders:
          type: integer