# lib/api-zod/src/generated/types/corporateSection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oxymed Medikal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CorporateSection {
  id: number;
  sectionKey: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}