# lib/api-zod/src/generated/types/mediaListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oxymed Medikal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaFile } from './mediaFile';

export interface MediaListResponse {
  items: MediaFile[];
  total: number;
}